Quick 'python -m pygame.docs', more testing
