--- docs/util.py
@@ -1,31 +1,0 @@
-import os
-import webbrowser
-from urllib.parse import urlunparse, quote
-
-
-def _iterpath(path):
-    path, last = os.path.split(path)
-    if last:
-        for p in _iterpath(path):
-            yield p
-        yield last
-
-# for test suite to confirm pygame built with local docs
-def has_local_docs():
-    pkg_dir = os.path.dirname(os.path.abspath(__file__))
-    main_page = os.path.join(pkg_dir, 'generated', 'index.html')
-    return os.path.exists(main_page)
-
-# called by __main__.py
-def open_docs():
-    pkg_dir = os.path.dirname(os.path.abspath(__file__))
-    main_page = os.path.join(pkg_dir, 'generated', 'index.html')
-    if os.path.exists(main_page):
-        url_path = quote('/'.join(_iterpath(main_page)))
-        drive, rest = os.path.splitdrive(__file__)
-        if drive:
-            url_path = "%s/%s" % (drive, url_path)
-        url = urlunparse(('file', '', url_path, '', '', ''))
-    else:
-        url = "https://www.pygame.org/docs/"
-    webbrowser.open(url)
